[🚚change]: CreateRoom파일 분리, 타입 분리
